Update 0.5
Added more functionality to UGrabber::Grab() - it can now detach and wake the physics of an object (intented to be grabbed)
UMover::Tick() now returns the moved platform back to its original position if the trigger connected to it loses the acceptable (trigger) component.
Added a golden statue reward in the mausoleum in at the end of the crypt as well as a pot on the other end of the crypt. When the statue is removed from the stand a trap door blocks the exit until an unlock component is put back in. The idea is to put the pot in place of the statue and escape the crypt.